+++ src/components/GuideLinks/GuideLinks.type.ts
@@ -8,6 +8,7 @@
 export type GuideLinksType =  {
   links: {
     label: React.ReactNode;
-    url: string;
+    url?: string;
   }[];
+  textAlign?: 'left' | 'center' | 'right';
 };