+++ src/setupTests.ts
@@ -8,6 +8,11 @@ import '__mocks__/iconify-react';
 
 jest.mock('react-markdown');
 jest.mock('react-i18next', () => ({
-  useTranslation: () => ({ t: (key: string) => key }),
+  useTranslation: () => ({
+    t: (key: string) => key, 
+    i18n: {
+      language: undefined
+    } 
+  }),
 }));
 jest.mock('@fortawesome/react-fontawesome');