+++ changelog.md
@@ -1,13 +1,23 @@
 # Changelog of quotly-frontend
 All notable changes to this project will be documented in this file.
 
+## [0.6.0] - 2024-11-09
+### Additions
+- Added Landing Page for [#15](https://github.com/quotly-eu/quotly-frontend/issues/15)
+- Added styling templates util for easy access for sharing css rulesets which are not worth to be directly a component.
+
+### Changes
+- Updated useFetch Hook with RequestInit arg
+- Updated env App description
+- Updated GuideLinks to allow to center the items
+
 ## [0.5.0] - 2024-11-07
 ### Additions
 - Added ApiContext and useFetch Hook for [#12](https://github.com/quotly-eu/quotly-frontend/issues/12)
 - Added ApiResponse & User type declaration 
 
 ### Changes
-- Updated Button & Input Components
+- Updated Button, GuideLinks & Input Components
 - Updated index file with ApiContextProvider
 
 ## [0.4.0] - 2024-11-07